weibo-instant-view/requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
idna==2.9
pkg-resources==0.0.0
pycparser==2.19
python-telegram-bot==12.4.2
requests==2.23.0
six==1.14.0
soupsieve==2.0
tornado==6.0.3
urllib3==1.25.8
